Crate pallet_ethereum_xcm

Ethereum Xcm pallet

The Xcm Ethereum pallet is a bridge for Xcm Transact to Ethereum pallet

Re-exports

pub use self::pallet::*;

Modules

pallet

The pallet module in each FRAME pallet hosts the most important items needed to construct this pallet.

Structs

AccessListItem

EnsureXcmEthereumTransaction

LegacyTransactionMessage

Log

MessageProcessorWrapper

TransactionStatus

Enums

EthereumXcmTransaction

Xcm transact’s Ethereum transaction.

RawOrigin

Receipt

Transaction

TransactionAction

Traits

EnsureProxy

Ensure that a proxy between delegator and delegatee exists in order to deny or grant permission to do xcm-transact to transact_through_proxy.

XcmToEthereum

Functions

ensure_xcm_ethereum_transaction

Type Aliases

Block