Module pallet_ethereum_xcm::pallet::dispatchables

Auto-generated docs-only module listing all defined dispatchables for this pallet.

Warning: Doc-Only

Members of this module cannot be used directly and are only provided for documentation purposes. To see the real version of each dispatchable, look for them in Pallet or Call.

Functions

force_transact_as

Xcm Transact an Ethereum transaction, but allow to force the caller and create address. This call should be restricted (callable only by the runtime or governance). Weight: Gas limit plus the db reads involving the suspension and proxy checks

resume_ethereum_xcm_execution

Resumes all Ethereum executions from XCM.

suspend_ethereum_xcm_execution

Suspends all Ethereum executions from XCM.

transact

Xcm Transact an Ethereum transaction. Weight: Gas limit plus the db read involving the suspension check

transact_through_proxy

Xcm Transact an Ethereum transaction through proxy. Weight: Gas limit plus the db reads involving the suspension and proxy checks