pallet_ethereum_xcm::pallet

Trait Config

pub trait Config:
    Config<RuntimeEvent: From<Event<Self>>>
    + Config
    + Config<RuntimeEvent: From<Event<Self>>> {
    type InvalidEvmTransactionError: From<TransactionValidationError>;
    type ValidatedTransaction: ValidatedTransaction;
    type XcmEthereumOrigin: EnsureOrigin<Self::RuntimeOrigin, Success = H160>;
    type ReservedXcmpWeight: Get<Weight>;
    type EnsureProxy: EnsureProxy<<<Self as Config>::AccountProvider as AccountProvider>::AccountId>;
    type ControllerOrigin: EnsureOrigin<Self::RuntimeOrigin>;
    type ForceOrigin: EnsureOrigin<Self::RuntimeOrigin>;
}

Configuration trait of this pallet.

The main purpose of this trait is to act as an interface between this pallet and the runtime in which it is embedded in. A type, function, or constant in this trait is essentially left to be configured by the runtime that includes this pallet.

Consequently, a runtime that wants to include this pallet must implement this trait.

Required Associated Types

type InvalidEvmTransactionError: From<TransactionValidationError>

Invalid transaction error

type ValidatedTransaction: ValidatedTransaction

Handler for applying an already validated transaction

type XcmEthereumOrigin: EnsureOrigin<Self::RuntimeOrigin, Success = H160>

Origin for xcm transact

type ReservedXcmpWeight: Get<Weight>

Maximum Weight reserved for xcm in a block

type EnsureProxy: EnsureProxy<<<Self as Config>::AccountProvider as AccountProvider>::AccountId>

Ensure proxy

type ControllerOrigin: EnsureOrigin<Self::RuntimeOrigin>

The origin that is allowed to resume or suspend the XCM to Ethereum executions.

type ForceOrigin: EnsureOrigin<Self::RuntimeOrigin>

An origin that can submit a create tx type

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.