Trait pallet_ethereum_xcm::pallet::Config

pub trait Config: Config + Config + Config {
    type InvalidEvmTransactionError: From<InvalidEvmTransactionError>;
    type ValidatedTransaction: ValidatedTransaction;
    type XcmEthereumOrigin: EnsureOrigin<Self::Origin, Success = H160>;
    type ReservedXcmpWeight: Get<Weight>;
    type EnsureProxy: EnsureProxy<Self::AccountId>;
    type ControllerOrigin: EnsureOrigin<Self::Origin>;
}

Configuration trait of this pallet.

Implement this type for a runtime in order to customize this pallet.

Required Associated Types

type InvalidEvmTransactionError: From<InvalidEvmTransactionError>

Invalid transaction error

type ValidatedTransaction: ValidatedTransaction

Handler for applying an already validated transaction

type XcmEthereumOrigin: EnsureOrigin<Self::Origin, Success = H160>

Origin for xcm transact

type ReservedXcmpWeight: Get<Weight>

Maximum Weight reserved for xcm in a block

type EnsureProxy: EnsureProxy<Self::AccountId>

Ensure proxy

type ControllerOrigin: EnsureOrigin<Self::Origin>

The origin that is allowed to resume or suspend the XCM to Ethereum executions.